pyproject.toml
[tool.poetry]
name = "django-learning-platform"
version = "0.0.1"
description = ""
authors = ["hoopengo <hoopengo@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"

[tool.poetry.group.dev.dependencies]
flake8-comprehensions = "^3.10.1"
flake8-quotes = "^3.3.2"
flake8-string-format = "^0.3.0"
flake8-black = "^0.3.3"
flake8-commas = "^2.1.0"
flake8-bugbear = "^22.7.1"
flake8-polyfill = "^1.0.2"
black = "^23.1.0"
flake8 = "^6.0.0"
pytest = "^6.2.5"
isort = "^5.10.1"
pep8-naming = "^0.13.0"
pre-commit = "^2.19.0"
autoflake = "^1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
